requirements.txt
pyspark==3.1.1
kafka-python
psycopg2-binary
pandas